parallelises the remaining serial functionality so that now virtually all of the code runs concurrently. This leads to a large improvement in both speed, and memory usage.
